rm functions for QRCode
