+++ src/QRDecoders.jl
@@ -1,12 +1,8 @@
 module QRDecoders
 
-## import functions from QRCoder.jl
-using QRCode
-using QRCode.Polynomial
-import QRCode: formatinfo
-import QRCode.Polynomial: logtable, antilogtable, mult, generator, lead, tail!, init!
+export Poly
 
-## add functions to QRCode
-include("qrcoder.jl")
+using QRCode
+import QRCode: Poly
 
 end
