+++ src/qrinfo.jl
@@ -1,0 +1,68 @@
+# infomation of QR-code
+
+## Hamming distance
+"""
+    hamming_weight(x::Poly)
+
+Calculate the Hamming weight of a polynomial.
+"""
+hamming_weight(x::Poly) = hamming_weight(x.coeff)
+hamming_weight(x::AbstractVector) = count(!iszero, x)
+hamming_weight(x::Integer) = hamming_weight(digits(x; base=2))
+
+"""
+    hamming_distance(x::Poly, y::Poly)
+
+Calculate the Hamming distance between two polynomials.
+"""
+hamming_distance(x::Poly, y::Poly) = hamming_distance(x.coeff, y.coeff)
+hamming_distance(x::AbstractVector, y::AbstractVector) = hamming_weight(x - y)
+hamming_distance(x::Integer, y::Integer) = hamming_weight(x ⊻ y)
+
+## Functions about encoding & decoding
+"""
+    qrformat(fmt::Int)
+
+Generate standard format infomation (format + error correction + mask).
+"""
+function qrformat(fmt::Int)
+    err = fmt << 10
+    g = 0x537 ## generator polynomial(= 0b10100110111 in binary)
+    for i in 4:-1:0
+        if !iszero(err & (1 << (i + 10)))
+            err ⊻= g << i
+        end
+    end
+    fmt << 10 ⊻ err ⊻ 0x5412 ## mask(= 0b101010000010010 in binary)
+end
+qrformat(fmt::Integer) = qrformat(Int(fmt)) ## to avoid integer overflow
+
+"""
+Bit data of information of quality.
+"""
+quality2binary = Dict(
+    Low() => 0b01,
+    Medium() => 0b00,
+    Quartile() => 0b11,
+    High() => 0b10)
+binary2quality = Dict(val=>key for (key, val) in quality2binary)
+
+
+"""
+    qrdecode_format(fmt::Int)::Int
+
+Decode format information.
+"""
+function qrdecode_format(fmt::Int)::Int
+    best_fmt, best_dist = -1, 15
+    for test_fmt in 0:31
+        test_code = qrformat(test_fmt)
+        test_dist = hamming_weight(fmt ⊻ test_code)
+        if test_dist < best_dist
+            best_dist = test_dist
+            best_fmt = test_fmt
+        end
+    end
+    ## best_dist > 5 && return -1 # too many errores
+    return best_fmt
+end
